(PC-22877)[API] refactor: move venue and offerer deletion methods
