+++ api/src/pcapi/admin/custom_views/offerer_view.py
@@ -16,12 +16,12 @@ from pcapi.core.external import zendesk_sell
 from pcapi.core.external.attributes.api import update_external_pro
 from pcapi.core.history import api as history_api
 from pcapi.core.history import models as history_models
+import pcapi.core.offerers.api as offerers_api
 import pcapi.core.offerers.exceptions as offerers_exceptions
 from pcapi.core.offerers.models import Offerer
 from pcapi.core.offerers.models import OffererTag
 import pcapi.core.offerers.repository as offerers_repository
 from pcapi.models.validation_status_mixin import ValidationStatus
-from pcapi.scripts.offerer.delete_cascade_offerer_by_id import delete_cascade_offerer_by_id
 from pcapi.utils.urls import build_pc_pro_offerer_link
 
 
@@ -118,7 +118,7 @@ class OffererView(BaseAdminView):
         emails = offerers_repository.get_emails_by_offerer(offerer)
 
         try:
-            delete_cascade_offerer_by_id(offerer.id)
+            offerers_api.delete_offerer(offerer.id)
         except offerers_exceptions.CannotDeleteOffererWithBookingsException:
             flash("Impossible d'effacer une structure juridique pour laquelle il existe des réservations.", "error")
             return False
